[anonymous-read] 벡터 평균 유사도 검색 추가

+++ domain/src/main/kotlin/com/techinsights/domain/repository/post/PostEmbeddingJpaRepository.kt
@@ -15,5 +15,5 @@ interface PostEmbeddingJpaRepository : JpaRepository<PostEmbedding, Long>{
   """,
     nativeQuery = true
   )
-  fun findSimilarPosts(targetVector: String, excludeIds: List<Long>, limit: Int): List<PostEmbedding>
+  fun findSimilarPosts(targetVector: String, excludeIds: List<Long>, limit: Long): List<PostEmbedding>
 }
