+++ domain/src/main/kotlin/com/techinsights/domain/repository/post/PostEmbeddingRepository.kt
@@ -4,5 +4,5 @@ import com.techinsights.domain.dto.embedding.PostEmbeddingDto
 
 interface PostEmbeddingRepository {
   fun findByPostIdIn(postIds: List<Long>): List<PostEmbeddingDto>
-  fun findSimilarPosts(targetVector: String, excludeIds: List<Long>, limit: Int): List<PostEmbeddingDto>
+  fun findSimilarPosts(targetVector: String, excludeIds: List<Long>, limit: Long): List<PostEmbeddingDto>
 }
